backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
redis==6.4.0
celery==5.5.3
python-dotenv==1.1.1
Pillow==10.4.0
openai==1.51.0
anthropic==0.25.0
requests==2.31.0
python-decouple==3.8
gunicorn==21.2.0
uvicorn==0.27.1
channels==4.0.0
daphne==4.0.0
